chore(storybook): split deck directive stories
